refactor to ArrowRoundingPolicy and BufferAllocatorUtil
