+++ fluss-common/src/main/java/org/apache/fluss/record/LogRecordReadContext.java
@@ -25,6 +25,7 @@ import org.apache.fluss.metadata.TableInfo;
 import org.apache.fluss.row.InternalRow;
 import org.apache.fluss.row.InternalRow.FieldGetter;
 import org.apache.fluss.row.ProjectedRow;
+import org.apache.fluss.row.arrow.memory.BufferAllocatorUtil;
 import org.apache.fluss.shaded.arrow.org.apache.arrow.memory.BufferAllocator;
 import org.apache.fluss.shaded.arrow.org.apache.arrow.vector.VectorSchemaRoot;
 import org.apache.fluss.types.DataType;
@@ -114,7 +115,7 @@ public class LogRecordReadContext implements LogRecordBatch.ReadContext, AutoClo
             boolean projectionPushDowned,
             SchemaGetter schemaGetter) {
         // TODO: use a more reasonable memory limit
-        BufferAllocator allocator = FlussRoundingPolicy.createBufferAllocator();
+        BufferAllocator allocator = BufferAllocatorUtil.createBufferAllocator();
         FieldGetter[] fieldGetters = buildProjectedFieldGetters(dataRowType, selectedFields);
         return new LogRecordReadContext(
                 LogFormat.ARROW,
